Add SingleViewIterator, massively reduce code duplication!
