+++ fabric-transfer-api-v1/src/testmod/java/net/fabricmc/fabric/test/transfer/fluid/CreativeFluidStorage.java
@@ -17,7 +17,6 @@
 package net.fabricmc.fabric.test.transfer.fluid;
 
 import java.util.Iterator;
-import java.util.NoSuchElementException;
 
 import net.minecraft.fluid.Fluids;
 
@@ -25,6 +24,7 @@ import net.fabricmc.fabric.api.transfer.v1.fluid.FluidKey;
 import net.fabricmc.fabric.api.transfer.v1.fluid.FluidPreconditions;
 import net.fabricmc.fabric.api.transfer.v1.storage.StorageView;
 import net.fabricmc.fabric.api.transfer.v1.storage.base.ExtractionOnlyStorage;
+import net.fabricmc.fabric.api.transfer.v1.storage.base.SingleViewIterator;
 import net.fabricmc.fabric.api.transfer.v1.transaction.Transaction;
 
 public class CreativeFluidStorage implements ExtractionOnlyStorage<FluidKey>, StorageView<FluidKey> {
@@ -70,42 +70,11 @@ public class CreativeFluidStorage implements ExtractionOnlyStorage<FluidKey>, St
 
 	@Override
 	public Iterator<StorageView<FluidKey>> iterator(Transaction transaction) {
-		CreativeFluidIterator iterator = new CreativeFluidIterator();
-		transaction.addCloseCallback(iterator);
-		return iterator;
+		return SingleViewIterator.create(this, transaction);
 	}
 
 	@Override
 	public int getVersion() {
 		return 0;
-	}
-
-	private class CreativeFluidIterator implements Iterator<StorageView<FluidKey>>, Transaction.CloseCallback {
-		boolean open = true;
-		boolean hasNext = true;
-
-		@Override
-		public boolean hasNext() {
-			return open && hasNext && amount() > 0;
-		}
-
-		@Override
-		public StorageView<FluidKey> next() {
-			if (!open) {
-				throw new NoSuchElementException("The transaction for this iterator was closed.");
-			}
-
-			if (!hasNext()) {
-				throw new NoSuchElementException();
-			}
-
-			hasNext = false;
-			return CreativeFluidStorage.this;
-		}
-
-		@Override
-		public void onClose(Transaction transaction, Transaction.Result result) {
-			open = false;
-		}
 	}
 }
